Rename python test.py to test_basic.py to parallel C test name.  See issue #922
