+++ bindings/python/Makefile
@@ -64,7 +64,7 @@ clean:
 	rm -f MANIFEST
 
 
-TESTS = test.py test_detail.py test_arm.py test_arm64.py test_mips.py test_ppc.py
+TESTS = test_basic.py test_detail.py test_arm.py test_arm64.py test_mips.py test_ppc.py
 TESTS += test_sparc.py test_systemz.py test_x86.py test_xcore.py test_skipdata.py
 check:
 	@for t in $(TESTS); do \
